move seq code into a module
